+++ app/lib/shared/versions.dart
@@ -21,9 +21,9 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// Make sure that at least two versions are kept here as the next candidates
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
-  '2021.05.11', // The current [runtimeVersion].
+  '2021.05.17', // The current [runtimeVersion].
+  '2021.05.11',
   '2021.04.27',
-  '2021.04.06',
 ];
 
 /// Represents a combined version of the overall toolchain and processing,
@@ -53,10 +53,10 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.12.0';
-final String toolStableDartSdkVersion = '2.12.4';
+final String toolStableDartSdkVersion = '2.13.0';
 final String toolStableFlutterSdkVersion = '2.0.6';
-final String toolPreviewDartSdkVersion = '2.13.0-211.14.beta';
-final String toolPreviewFlutterSdkVersion = '2.2.0-10.2.pre';
+final String toolPreviewDartSdkVersion = '2.13.0';
+final String toolPreviewFlutterSdkVersion = '2.2.0-10.3.pre';
 
 // Value comes from package:pana.
 final String panaVersion = pana.packageVersion;